+++ frontend/src/pwa/badge.js
